:tractor: Extract UI to its package, fix some sizing issues
