+++ pkg/ui/utils.go
@@ -1,4 +1,4 @@
-package main
+package ui
 
 import (
 	"fmt"
